# static/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "The Unchained MHacks API",
        "version": "0.0.1",
        "title": "MHacks",
        "termsOfService": "https://mhacks.org",
        "contact": {
            "email": "me@antoniomika.me"
        },
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://github.com/mhacks/mhacks-web/blob/master/LICENSE"
        }
    },
    "servers": [
        {
            "url": "https://mhacks.org/v1",
            "description": "Production"
        },
        {
            "url": "https://staging.mhacks.org/v1",
            "description": "Staging"
        },
        {
            "url": "http://localhost:3000/v1",
            "description": "Development"
        }
    ],
...
    "paths": {
...
        "/v1/push": {
            "patch": {
                "tags": [
                    "Push"
                ],
                "summary": "",
                "description": "",
                "operationId": "patchPush",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Push"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "push": {
                                            "$ref": "#/components/schemas/Push"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Push"
                ],
                "summary": "",
                "description": "",
                "operationId": "putPush",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Push"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "push": {
                                            "$ref": "#/components/schemas/Push"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Push"
                ],
                "summary": "",
                "description": "",
                "operationId": "postPush",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Push"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "push": {
                                            "$ref": "#/components/schemas/Push"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "Push"
                ],
                "summary": "",
                "description": "",
                "operationId": "getPush",
                "responses": {
                    "200": {
                        "description": "Okay",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "description": "Whether or not the request was successful"
                                        },
                                        "push": {
                                            "$ref": "#/components/schemas/Push"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error (Unknown Error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "schemas": {
...
            "Push": {
                "type": "object",
                "properties": {}
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}